requirements.txt
torch
pandas
transformers
datasets
tqdm
kiwipiepy
bertopic
liqfit
SentencePiece
jupyter
soynlp
python-mecab-ko
kss
flash-attn